dash/app/requirements.txt
fastapi
uvicorn[standard]
requests
pyyaml
